Fix derive Equivalence for arrays + formatting
Allows the derive macro to be used on static arrays with non `{integer}`
type constants, Specifically, `usize`. I include a test to prevent
regression.

I made 2 cosmetic changes in collective.rs and examples/struct.rs to
allow `rsmpi` to compile without warnings.

+++ mpi-derive/src/lib.rs
@@ -43,7 +43,7 @@ fn equivalence_for_tuple_field(type_tuple: &syn::TypeTuple) -> TokenStream2 {
 fn equivalence_for_array_field(type_array: &syn::TypeArray) -> TokenStream2 {
     let ty = equivalence_for_type(&type_array.elem);
     let len = &type_array.len;
-    quote! { &::mpi::datatype::UncommittedUserDatatype::contiguous(#len, &#ty) }
+    quote! { &::mpi::datatype::UncommittedUserDatatype::contiguous(#len as i32, &#ty) }
 }
 
 fn equivalence_for_type(ty: &syn::Type) -> TokenStream2 {
